Added test environ that supports simplestreams.HasRegion
